Separate js file in modules
