+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
     <title>Awesome books</title>
 </head>
 <body>
@@ -57,6 +57,6 @@
   </footer>
 
 
-  <script src="./script.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
