+++ modules/book.js
@@ -1,0 +1,6 @@
+export class Book {
+  constructor(name, author) {
+    this.name = name;
+    this.author = author;
+  }
+}
