+++ modules/clock.js
@@ -1,0 +1,5 @@
+export function displayClock() {
+    document.getElementById('dateandtime').innerHTML = Date();
+    setTimeout(displayClock, 1000);
+}
+
