+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "awesome-books",
+  "version": "1.0.0",
+  "description": "> \"Awesome books\" is a simple website that displays a list of books and allows you to add and remove books from that list. For building this application, we used local storage to save the data of the books and we give it some interactivity using JavaScript. We also used a medium-fidelity wireframe to build the UI.",
+  "main": "script.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/marurevi/Awesome-books.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/marurevi/Awesome-books/issues"
+  },
+  "homepage": "https://github.com/marurevi/Awesome-books#readme"
+}
